Add "./manage.py landfill" command.
Could very likely be nicer but this takes the commands from
the wiki and makes it a single command. Has two optional
arguments --no-notice which runs the command from the wiki to
remove the landfill notice and --no-download which uses an
already downloaded file instead of downloading it.

cleaned up install_landfill command

notice regarding gzcat/*nix

removed some redundancy in install_landfill