Fixed issues with schema metadata on add-on pages (bug 709643)
Chrome Web Store had a description on each extension that they host, so
we should have that as well. This may be why we don't get rich snippets
in search results ("description" is an attribute of the "Thing" schema,
which is the root of all object schemas). Also changed from
SoftwareApplication to WebApplication, as this is what Google lists all
of its browser extensions as.