Fixed app lookup search for multi-word queries (bug 767571)
Fuzzy searches use Levenshtein distance and don't tokenize the input. So
in the case of this bug, searching for "mozilla marketplace" tried to
compare the edit distance of "mozilla marketplace" to "mozilla" (and
"marketplace" separately). But the edit distance of both of these was
more than the default minimum of smallest word in the query string
("mozilla" = 7 words). So it assumed these were too far apart.

This patch adds a text query so either should match for a result.