Unmark char counters as errors after form validation (bug 769244)
After form validation, character counters are wrapped in a block with
the class "error". Because of this, even after the field has been
fixed to the correct length, the character counter is highlighted in
red. This marks the field as not being an error after the length has
been adjusted down.

It's possible that the text field also has other errors, but it's
assumed that if the user has modified the field and it's within the
length requirements, it shouldn't continue to be flagged, anyway, so
that case doesn't need to be dealt with.