Bug 792795 - Call "dibs" on installing persona handlers
Unfortunately, both (MKT)login.js and (AMO)browserid_support.js files
are included on some templates. This can lead to bad behaviors as the
wrong handler may be called for certain events.

The patch has a handler set a 'pers-handle' data handle on the page body
to indicate that it is handling the persona transactions.

added ; after returns