Fix for self-xss, double-escape (bug 711833)
First, use `text()` to get the text value of the review. We don't want
HTML entities to ever be seen by the user. Pulling `html()` and
replacing line breaks is an enormous hack.

Second, escape the HTML that we're injecting back into the page to
prevent the user from XSSing themselves.