Review replies should bounce to list (bug 807283)
Previously, there was code that did pseudo-POST hijacking. Not
anymore (because we have actual POST hijacking), but that means that
all of the wonky code that depended on it is going to float to the
surface.

I'm probably going to rip out a lot of code from
media/js/mkt/ratings.js later, but this should solve the issue of the
browser not leaving the modal dialog.