Fighting w/Jenkins to fix recaptcha test
Jenkins is configured without a recaptcha key (did not know this).
This means that recaptcha won't be displayed, and an error won't be
raised. We need to mock in a (dummy) key to turn RC on, but it will
always throw an exception, so that's just fine. We want to make sure
that it shows up and it blocks zombie submissions.