Better link regex for link detection (bug 785326)
This change does a few things:

1. Removes super slow regex code that matches exact hostnames
2. Relaxes regex to match patterns like 'example.tld'
3. Re-orders regex to try to match for protocols and IPs before
   more generic patterns.
4. Matches more protocols (we don't want `skype://` links either)

The primary change removes the strict structure of compliant
hostnames and replaces it with a pattern that will match anything
that looks *similar* to a hostname. This means things like
'foo.x.yz' will set it off, but that's probably suspicious enough
to warrant a flag anyway.

No unit tests should need to be updated, because this should have
roughly the same behavior as the existing implementation, sans the
headaches and heartbreaks.