Fix for Chrome's broken Vary support (bug 774899)
When we tell Chrome to fetch a page, we want to make sure that we
don't hit the cache if the user has changed his language or region
preference. To do this, we "bust" the cache by including the region
and the locale (if present) in the query parameters.

What this means is that Chrome will never try to hit the cache after
the locale or region changes.