Bump clean-css, uglify-js versions
clean-css is used by jingo-minify, which executes it as a subprocess. The CLI arguments have not changed between 0.x and 2.0.x, so this should be safe. We should notice a distinct size win in our asset bundles.

uglify-js is the same: the arguments have not changed and the latest version(s) offer bug fixes, efficiency improvements, and minor size wins.