treat "worldwide" as "restofworld"
When looking up a region treat "worldwide" as "restofworld", please use
the REGION_LOOKUP dict instead of REGIONS_DICT for looking up regions by
slug.
Handle bad regions in region detail API.
The submission API docs said that you can set the regions with a country
code but the regions field in the serializer is set to `read_only`.
Close #1649.