Remove obsolete mobile documentation links
I updated some of the mobile links to point to up-to-date resources on MDN, but I removed the links that don't have a modern equivalent. I'm taking note of those links to create new MDN articles, but in the meantime, I think it's better to point to fewer, yet accurate, documents.