Remove non-documented legacy API parts (fix #1715, fix #1721)
Everything behind authentication in the old API is removed in this
commit. In addition, everything behind api version "2" is also
removed as it is undocumented. (Firefox uses version "1.5", the
newer stuff we build uses version "3". AFAIK version "2" was only
used for things like the add-on builder, which no longer exists)

This also removes piston, since it was only useful for those
undocumented APIs.

This does *not* switch the remaining parts of the legacy API to
DRF yet.