Bugzilla-1254143: Switch to pip hash checking mode
Added hashes for modules whose versions are pinned with "==". Need to
break the existing requirement files into 2 types: with and without
hashes. This is because according to the doc,

> Hash verification is an all-or-nothing proposition. Specifying a
> --hash against any requirement not only checks that hash but also
> activates a global hash-checking mode.

For example, for those modules whose versions are not pinned with "==",
or those that are installed from VCS directly, if they're mixed in the
same file with modules that use "--hash" option, the install fails with
errors similar to the following:

> In --require-hashes mode, all requirements must have their versions
> pinned with ==. These do not:
>     simplejson>=3.0.5 from https://pyrepo.stage.mozaws.net/olympia/simplejson-3.8.1.tar.gz
>     (from -r compiled_without_hash.txt (line 16))

or

> The editable requirement django-queryset-transform from
> git+https://github.com/jbalogh/django-queryset-transform@a1ba6ae41bd86f5bb9ff66fb56614e0fafe6e022#egg=django-queryset-transform
> (from -r compiled_without_hash.txt (line 17)) cannot be
> installed when requiring hashes, because there is no single file to
> hash.