Remove package.json related extraction code, ignore it everywhere.
This doesn't remove everything related to package.json though since it
appears that it's not used but is part of many add-ons and XPI files
so I left the version increment on resigning and the FileViewer code
untouched.

Also the amo-validator continues to validate package.json related
code as long as we accept "new style" jetpack add-ons.

Fixes #2460, refs #2564, refs #2438