Properly default to our default AppVersions even if applications[/gecko] is omitted.
Fixes #2586 and probably a few others.

WebExtensions are allowed to omit `applications[/gecko]` and we
previously skipped defaulting to any `AppVersion` once this is
not defined. That resulted in none of our plattforms being selectable.

Since WebExtensions can skip defining `applications[/gecko]` (e.g from
the Chrome Store) we should ignore this too.