Swap jobs order in travis to make builds faster
'es' is usually 2 minutes faster than 'main', and the last job to
run usually starts later than the rest because it needs to wait
for one of the first jobs to be finished to start. So that last
job needs to be the fastest job that we have to speed the whole
build up.