Remove *_from_hive commands, they were never used in production.
Those were intended to replace the system we have now where we run
the queries on a separate server, generating a file that we then
copy to addons-server servers to parse it.

The replacements are much more sane than the system we have in place
now but we don't have the time to investigate the weird issues that
lead the previous team to never switch to using those scripts in
production.