Refactor the way categories are attached to Addons using StaticCategory
This makes all_categories, app_categories, get_category coherent,
using the same source, so that it's easier to then override when
needed. app_categories is changed to return a dict, because it's
more useful that way.

This removes the need for the special attach_categories() function
that the add-on indexer was using, as we now load all categories
for each add-on instead of just the first one.

The final (big!) change is to use StaticCategory instances instead of
Category for these properties, because we have all the relevant
information in the constants.