Replace Review.objects.valid with Review.without_deleted, removing a useless JOIN
The default objects manager does an extra JOIN on reply_to, to hide replies to
deleted reviews, but when we have already decided to not care about replies,
this join is useless. The .valid() method, on top of having a weird name,
could not easily change this behavior because it's too late, so this commit
renames to 'without_replies' and moves it to a manager instead.