Fix bugzilla-1305466, wrong file ownership
Environment variable UWSGI_GID makes uwsgi override the command line
--gid option, causing files written by the containers having wrong group
ownership of "uwsgi".

We don't need the uwsgi group specifically defined inside containers
because the containers are going to be run as olympia user and olympia
group anyways.

This change also requires that puppet-config to change the group ownership
of some shared directories such as /var/run/uwsgi to be "olympia".