Return browserVersion '' by default instead of 0
If detection succeeds, browserVersion is a string.
If it fails (e.g. because the user agent does not match any of the
supported user agents), then browserVersion is 0.

The problem with this is that code that assumes browserVersion to be a
string throws. For example, the `VersionCompare.compareVersions` method
assumes that the input is a string and calls `split` on its parameter.
When used with `z.browserVersion`, it throws in unsupported browsers.
For example, when a Firefox add-on listing is viewed in Chrome, the user
reviews do not appear because of the above error.

An empty string ('') is falsey just like `0`, so if there is any code
that uses `!z.browserVersion`, it still works.