Fix the lazy choices labels for DistributionChoiceForm to stay lazy.
'mark_lazy' doesn't keep objects lazy unfortunately, so the fix
introduced in 95de533a02f55a708d23e047242be008dee44355 only fixed
part of the problems.

This time we'll do it right. And we have tests!

What this and 95de533 are fixing: The problem here is that marking a
string with ugettext and not ugettext_lazy instanly translates them on
the first request. So what we saw on production was a different label at
https://addons.mozilla.org/en-US/developers/addon/submit/distribution
depending on what instance we were hitting and what user hit that
instance first (or more, with what locale it got initialized because of
the user who hit the instance first after deployment).

Keeping strings lazy avoids this problem and translates them as late as
possible, and only during our request-response cycle and not during
import time.

Now that's what 95de533 did and this pull request fixes the fact that we
actually called mark_safe on the lazy string instances which instantly
translated them again. See django/django@0775ab2 for a few more details.

* Implement workaround to make sure labels get translated on run-time not import-time.