Make reviewer tools query simpler, don't do a subselect for the latest version
Since we're filtering on file status, the version we get with a simple JOIN
should always be the right one, except for the few add-ons which are in a
weird state because they have been modified manually.

Note that the unlisted add-ons queue can't get that fix, because it's not
possible to use that trick when all versions are public. The unlisted add-on
count is removed from the menus in this commit to avoid doing the nasty
query as much as possible.

Flags on the review page are also changed to avoid calling the queue
query as well.