Make references to 2 env variables consistent across envs
This is to make the reference to EMAIL_QA_ALLOW_LIST and EMAIL_DENY_LIST
consistent across all environments. This helps mitigate the risk of the
same set of configs works on environment A but not B because a variable
is reference in B, but not A, or vice versa.