Remove obsolete code dealing with mirrors
We haven't been mirroring add-on files to external servers for a
long time: currently we just have files in 2 different locations
depending on whether they are disabled or not, and the CDN
mirrors the public ones.

All code dealing with mirrors is therefore obsolete.
{File}.file_path is the same thing as {File}.mirror_file_path,
making the latter redundant and copy_to_mirror(),
as well as amo.MIRROR_STATUSES, useless.

Fix #3887