Verify source and target folders after file-browser extraction.
Fixes #4316, or at least tries to.

The idea behind this patch is that we have a potential issue with Amazon
EFS in a way that it does writes partially asynchronous. Usually these
are fast enough but sometimes we run into a race condition that the
written / copied files from an extraction aren't properly read right
after calling shutil.copytree().

This patch tries to fix this by comparing the original list of files we
get during the extraction with what we have on our target system, EFS.

If the file-list isn't identical we go through the list and call `fsync`
for every file and folder in there. This should flush all buffers and
make sure that things are persistent before we are re-reading the
file-list.