Make sure file-viewer and it's cleanup cron are not relying on atime.
* Make sure that the file-viewer is extracting to a sub-folder with month-day in it to avoid cleanup crons trample on it
* Update cleanup_extracted_file to avoid reading any file metadata.
* Add a simple test that verifies it's functionality.

Hopefully fixes #4441

The idea behind this is the following: Amazon EFS, or more our mounts of
it are mounted with the `noatime` option so it's silly to actually rely
on it.

With this patch we now create a new sub-directory for every day and
extract into that, cleanup crons don't touch todays folders, ever. This
should improve reliance on existing folders and avoid crons trampling
into reading a file-list and pushing it into cache.