Implement auto-approval command prototype
This adds an 'auto_approve' management command, which fetches versions
from the updates queue, and try to see if they pass the criteria for
auto-approval. It stores the result of the tests for each version in
a model for future analysis.

At the moment actual approval action is not implemented yet.