Insert the exact number of previously approved webextensions in addon approvals counter table
Previous migration was just setting the counter to 1 for each public add-on. This sets it to
the number of approved, non-deleted, listed webextensions versions, making the initial counter
values more accurate for each add-on.

It ignores non-webextensions, in order to not fill the table with useless data but also to
not end up auto-approving an add-on until a reviewer has seen at least its first webextension
(in the case of an add-on that migrated from legacy to webext)