Remove bogus primary property from paypal API call
The docs indicate:
Set to true to indicate a chained payment; only one
receiver can be a primary receiver. Omit this field,
or set it to false for simple and parallel payments.

I suspect it was just ignored when we were passing "TRUE", and
it should be false or just omitted. We don't provide a secondary
recipient anyway.