Prevent dynamic fields in reviewer tools from being shown for the wrong action
fields with data-value="reject_multiple_versions" were shown for the "reject"
action because the value started with "reject"... Added a separator at the
end of every value list, and made sure to always include the separator when
searching, guaranteeing an exact match.

Also limit reject_multiple_versions to listed, public add-ons and make
sure the same version is not shown multiple times if it has multiple files
while we're at it.