Simplify Add-on type constants: remove ADDON_ANY except where it's actually used
ADDON_ANY is not an actual type choice: there are no addons with that
type, it's only a special value to be handled by search forms.

The only 2 places where we display a type widget selector are editors
tools and admin search. The former does need ADDON_ANY, the latter
does not because the admin already has a "All" option when displaying
list_filter choices.

Everywhere else on the site, we don't display a type widget when
searching, so we don't need to support the ADDON_ANY choice.