Fix includes and excludes for ElasticSearch 5, add tests.
There's a `TODO` in one of the tests since apparently ElasticSearch
still includes a few fields in _source even though we set them in
`excludes`. I don't know why yet, or wether this is a ES bug or not.

I'll follow up on this in a separate issue since the ones still in there
aren't as important as the l10n fields that are properly excluded.

Fixes #6060