Unify disabled buttons behavior across the site
This commit contains multiple small changes that go together:

Some pages are not "impala"-ed and therefore don't get the impala
and restyle styles, so they weren't getting the pointer-events: none
rule that disable the buttons. Fixing those pages would require a
lot more work, so the impala button styles have been added to the
base css bundle instead.

In addition, the restyle was setting pointer-events: all on disabled
buttons, but since we are no longer showing the popup on those buttons
we can remove that.

It was also possible in some pages to get the wrong style for a
a.button element with an href that you had already visited, because
the restyle wasn't targeting visited links, so this is changed as
well. Since doing this added more rules that previously weren't
always applied, the background position and color of the concealed
button also had to be specified.

Finally, there was an extra case where the more-versions link was
missing the newline that separates it from the compatibility info,
this was fixed.