Avoid parsing add-on too much too early
check=False didn't prevent the relevant extractor to parse apps,
which in turn could raise a ValidationError, which meant we didn't
know whether the extension was a webextension or not and ended up
calling the legacy validator in some cases...

The `check` parameter is replaced by the more explicit `minimal` in
this refactor, and the methods modified to only return the minimal
set of data that we need when that parameter is True. This should
ensure we only raise ValidationError for truly unavoidable errors
such as unreadable zip files.