Fix featured extensions "jumping around" as you hover them.
Though the main fix is to remove the width/height: auto, removing
the compatibility messages also help prevent potential display issues
because it makes the box smaller. A user on Chrome probably does not
care much about which version of Firefox or which platforms the
add-ons are compatible with anyway.