Clarify requirements for add-on version list
I'm bad at Sphinx docs so apologies if this is the wrong format, but I was reading
the API docs and found this section a bit hard to parse. There are a lot of commas,
"and"s, and "or"s. I feel like this makes it a bit easier to grok.