Check that the response is a 200 before looking at the content in reviewer tests
This is because otherwise some tests might be completely wrong, if
they are testing for the *absence* of something, and the view actually
returns a 403 because the test has not been set up properly.

To make sure we don't end up doing that by accident in the future when
refactoring tests, That pattern is applied to (almost) all the tests.