Randomize unlisted add-on slugs created via the API.
Fixes #5955

This also simplifies `clean_slug` a bit by using a set operation to
find free numbers instead of a for loop that doesn't cover the whole
range.

We only set a random slug if none is given, otherwise we rely on postfix
increments.

This patch also removes possibility of exposing add-on ids or even the
model name as default slugs.