Use `confirmed` field to calculate size of code changes and auto-approved queue
Now that this field exist and is properly populated it allows us to simplify
the queryset to generate the queue and also calculate size of code changes
more accurately.

Also include a command to recalculate post-review weights, we'll need to
run it to benefit from the new size of code changes formula.