Change internal auth API token name to be able to clear it completely
Sending multiple Set-Cookie headers with the same name is not allowed
by the spec, even if they have a distinct domain attribute, so we could
not clear the cookie on logout reliably, depending on how the user
logged in. Only solution is to change the name, so that it will be
correct from now on.