Align theme update JSON with theme pages on AMO
For the same theme data, the theme install page on AMO provides `"#"` as accent/textcolor strings, if those properties are not present. The fact that the update service then provides `null` is breaking themes in Firefox 60 and later. It seems sensible to align the two views of what is essentially the same data.