Avoid server error on duplicate public slugs.
See https://sentry.prod.mozaws.net/operations/olympia-prod/issues/2870827/ - and that is a relict from the chrome add-ons -> firefox conversion stuff that was going on end of last year where tons and tons of duplicate add-ons were created.

We fixed that for unlisted add-ons but only fixed it for new, unlisted, uploads and never converted older unlisted add-ons to new slugs. So we have to increase the amount of public duplicate slugs we allow to avoid server errors. Unlisted slugs are always random so this should fix the issue, mostly.

Fixes #7990