Remove features from collections that won't be ported to the new frontend
This removes the following features from collections:
- Ability to browse all public collections and sort through them
- Collection feeds
- Collection admin edition (the functionality is kept but instead
  moved to the django admin)
- Collection contributors (instead, just allow people with
  Collections:Contribute to edit add-ons of the featured themes
  collection.

This allows us to get rid of some long-standing performance issues with
the database queries associated with those features, improving overall
stability of the site.

Database migration removing collections_users table will be added
in a later commit to keep backwards-compatibility when this one is
pushed to production.