Backport cache.get_or_set from Django 1.11.
Part of our upgrade to Django 1.11, this is backporting functionallity
we'd like to use after the upgrade already to our current code so that
we can find potentially drawbags early.

This also reduces our reliance on cache-machine which helps removing it.

This also few smaller optimizations along the way:

 * denied-name blocklists were cached language-dependend although they're not, this now increases the chance of cache-hits instead of misses
 * the django cache.get_or_set implementation fixes potential
 race-conditions between multiple requests setting data

Fixes #8529