Completely remove traces of caching.backends settings.
Most of this work landed already in
6a8fc341daaa2754d7ccf74ef44ea49e94f9f5f9 and
91605a4c1fa679f8563e1434a0145ffc1b607beb so this is just a follow-up to
remove all other traces.

The removed comment is no longer true, Django received all the necessary
fixes for LocMemCache. On the other hand, LoMemCache is no longer used
for a regular local setup because we are using memcached and try to be
as close to production as possible so this won't change anytime
soon.