Add a custom overwrite for update_product_details to support offline docker builds.
The problem is that the upstream update_product_details command only
disables a few checks but not all of them. We don't run a MySQL database
while building the docker image and when Django runs db-field checks
(which it does by default since Django 1.10) it requires a working MySQL
connection.

Fies #8822