Don't save the add-on instance unless the form is valid.
The problem here is that form validation already sets the add-on data
(e.g name etc) during validation so that, in case you are saving that
instance we end up having an empty add-on name. The previous
`form.is_valid` (for the addon form) is correctly returning `False` so
we really just shouldn't issue a save if we know the instance is
invalid.

I added a mental note and will open an issue to check what changed from
1.8 - 1.11 that `addon.name` receives the newly empty valie while form
validation. But I couldn't find any re-use of that pattern so I'm
confident that this will fix issues like that.

Fixes #8919