Make discovery pane content come from the database instead of being hardcoded
This uses DiscoveryItems as the source of the discovery pane content, using
a new position (and position_china for, well, China) field to control whether
or not an item should be shown and in which position.

This also uses translations generated in advance from the database, and it's
applied to both editorial content and recommendations.

Finally, this reverts 76745bfc59d51689e8ffc8dd181c09a29061d0c8 as it's no
longer needed, since it will be possible for QA to just change the content
through the admin if they want to test a particular scenario.