Make discovery pane content come from the database instead of being hardcoded
This uses `DiscoveryItems` as the source of the discovery pane content,
using a new `position` (and `position_china` for, well, China) field to
control whether or not an item should be shown and in which position.
To ensure a smooth transition, the fields have been added ahead of time
in 6fb00edbca48f2b16c12aee1ab569601af1f5101

This also uses translations generated in advance from the database, and
it's applied to both editorial content and recommendations.

Finally, this reverts 76745bfc59d51689e8ffc8dd181c09a29061d0c8 as it's
no longer needed, since it will be possible for QA or devs to just change
the content through the admin if they want to test a particular scenario.