Replace platform support with plain application selection during submission
This pull request removes supported platform selection and replaces it
with selecting application explicitly. This way the developer can choose
between "Firefox" and "Firefox for Android" support and platforms are
always naively set to "PLATFORMS_ALL" since we want to get rid of
platforms eventually.

This cleans up most of the platform selection code that resides in devhub, views and templates since it's essentially unused now
This doesn't touch any of the platform handling in models since all of that has to continue working for legacy add-ons and WebExtensions that have different platforms selected
We are defaulting to "Firefox" (Desktop) compatibility with Android compatibility being optional

Regarding static themes, they are only compatible with Firefox Desktop currently which works seemlessly since we are defaulting to Firefox Desktop compatibility by default anyway and you can't select explicit Android compatibility for static themes.

That's why I removed all that "platforms to exclude" code and didn't implement a replacement for compatibility apps, any objections to that? We can do that if we have to, e.g if we decide to default to Firefox and Android support for regular WebExtensions.

Fixes #8752
Fixes #6067

This allows selecting support for an application directly during submission