Add notes for reviewers & disable auto-approval on mozilla signed add-ons (#9439)
Add notes for reviewers & disable auto-approval on mozilla signed add-ons

Note that the disabling of auto-approvals only happens on initial add-on submission: for subsequent versions the value is not automatically modified, leaving admins free to leave it on or off. The message for reviewers is per version so it's added for every version signed with mozilla internal cert.