Prepare elasticsearch index for locale-specific exact matches
This is done in a somewhat backwards-compatible way - this commit
adds a raw field but search/filters.py doesn't use it yet, so that
searches still work on the old index. One key difference with before
is that the non-locale aware fields will now only contain data in
the default locale instead of all locales.