Reverse 'Continue' and 'Cancel and Disable...' buttons.
Personally, the current order adds a lot of stress to the user to click
the correct button.

When you upload your file (very first screen) the Continue button is on
the very left site, all continuous screens switch it to be on the right
hand site. Even though the 'Cancel' button is red, it takes a bit of
time to parse which button you're going to press and ending up in a
disabled / incomplete state isn't really nice.

This changes this and fixes #7556 along the way.