Proposal: Reverse 'Continue' and 'Cancel and Disable...' buttons. (#10142)
Personally, the current order adds a lot of stress to the user to click
the correct button.

When you upload your file (very first screen) the Continue button is on
the very left side, all continuous screens switch it to be on the right
hand site. Even though the 'Cancel' button is red, it takes a bit of
time to parse which button you're going to press and ending up in a
disabled/incomplete state isn't really nice.

This changes that behaviour and fixes #7556 along the way.

![screenshot from 2018-12-03 13-07-10](https://user-images.githubusercontent.com/139033/49372994-c6111580-f6fc-11e8-9b41-d3cebc69771a.png)

cc @jvillalobos or @wagnerand for opinions on that.