Don't explicitly install npm, update Python to 2.7.15
Our docker builds were broken because of some inter-dependency weirdness
going on when you're installing npm and nodejs in one apt-get call. To
be honest I was too lazy to look up the actual root cause given the fix
is quite simple and npm is part of nodejs anyway.

This PR updates the Python version to 2.7.15 along the way.