Force pygit2 global search path to be set to HOME environment variable.
It's a bit problematic since there isn't a real good default value since
it depends on how the app is hosted, if `HOME` is set correctly and many
other circumstances. Let's enforce it to be the users home to make sure
we are accessing a directory that we have the permissions to.

Fixes #10320