Fix addons serializers tests for Python 3
Prevent infinite loop by not just looking for an empty string to
break out of the loop in copy_stored_filed(), but any falsy value
(crucially, b'' needs to end the loop since the file is opened in
binary mode). In addition, change all while 1: to while True: for
consistency.