Remove our own atomic decorator, run create_version_for_upload in 'read committed'
This is an attempt to fix our enourmous amounts of deadlocks. Now I
finally found a reasonable candidate.

devhub.tasks:create_version_for_upload is being wrapped in a custom
atomic decorator that sets the transaction level to 'SERIALIZABLE'. To
be fair, I don't think it needs that.

We switched to 'read committed' some time ago and I think it makes sense
to unify this here too.