Don't skip adding compatibility apps because of manifest incompatibility.
For example, if the user uploads an add-on that sets the key
browser_specific_settings and sets strict_min_version to 42.0 the
add-on is potentially invalid / unable to be used but we should't fail
to add compatibility just because of that.

Erroring because of schema errors is part of the addons-linter.

Fixes #10697